add tests for send-receive consistency
we have this bug that is caused by retries getting messages sent and received out of sync, so you send m1, m1, m1 because of retries, then next time you send m3 and call recv() and expect m3, but get m1.

This test checks that doesn't happen. But it needs to spin up a web socket server...