fix marian client so retries don't get out of sync
this fixes a bug where retries cause message order to be offset. That is, because of retries, you'd send m1 three times. Then you'd send m2, expecting the result to be related to m2, but you'd get m1 (or a corrected version thereof).

I think this fixes that.

It also has smarter logic for automatically setting timeout and retries if the user doesn't specify them on instantiation.

Finally, it lets the user pass a url directly, if they prefer that over auto-generating from host/port